+++ lib/netzke/base.rb
@@ -22,18 +22,32 @@ module Netzke
   # * Session-based persistence (see {Netzke::State})
   #
   # == Class-level configuration
-  # You can configure component classes in Rails Application, e.g.:
-  #
-  #     config.netzke.basepack.grid_panel.column_filters_available = false
-  #
-  # Optionally, when used outside of Rails, you can also set the values directly on Netzke::Core.config (the Engine does it for you):
-  #
-  #     Netzke::Core.config.netzke.basepack.grid_panel.column_filters_available = false
-  #
-  # If both default and overriding values are hashes, the default value gets deep-merged with the overriding value.
   #
   # Netzke::Base provides the following class-level configuration options:
   # * default_instance_config - a hash that will be used as default configuration for ALL of this component's instances.
+  #
+  # == Referring to JavaScript configuration methods from Ruby
+  #
+  # Netzke allows use Ruby symbols for referring to pre-defined pieces of configuration. Let's say for example, that a toolbar needs to nest a control more complex than a button (say, a date field), and a component should still make it possible to make it's presence and position in the toolbar configurable. We can implement it like this:
+  #
+  #     action :do_something
+  #
+  #     def configure(c)
+  #       super
+  #       c.tbar = [:do_something, :date_selector]
+  #     end
+  #
+  # While :do_something here is referring to a usual Netzke action, :date_selector is not declared in actions. If our JavaScript mixin file contains a method called `dateSelectorConfig`, it will be executed at the moment of configuring `tbar` at client side, and it's result, a config object, will substitute `date_selector`:
+  #
+  #     {
+  #       dateSelectorConfig: function(config){
+  #         return {
+  #           xtype: 'datefield'
+  #         }
+  #       }
+  #     }
+  #
+  # This doesn't necessarily have to be used in toolbars, but also in other places in config (i.e. layouts).
   class Base
     include Session
     include State