+++ test/core_test_app/app/components/component_with_prebuilt_toolbar_control/javascripts/component_with_prebuilt_toolbar_control.js
@@ -1,0 +1,12 @@
+{
+  prebuiltControlConfig: function(config){
+    return {
+      xtype: 'datefield',
+      listeners: {
+        select: function(){
+          this.netzkeParent.body.update('Hi');
+        }
+      }
+    }
+  }
+}